+++ src/reactpy_django/__init__.py
@@ -3,7 +3,6 @@ import contextlib
 import nest_asyncio
 
 from reactpy_django import (
-    checks,
     components,
     decorators,
     hooks,
@@ -12,14 +11,10 @@ from reactpy_django import (
     types,
     utils,
 )
-from reactpy_django.websocket.paths import (
-    REACTPY_WEBSOCKET_PATH,
-    REACTPY_WEBSOCKET_ROUTE,
-)
+from reactpy_django.websocket.paths import REACTPY_WEBSOCKET_ROUTE
 
 __version__ = "4.0.0"
 __all__ = [
-    "REACTPY_WEBSOCKET_PATH",
     "REACTPY_WEBSOCKET_ROUTE",
     "html",
     "hooks",
@@ -27,7 +22,6 @@ __all__ = [
     "decorators",
     "types",
     "utils",
-    "checks",
     "router",
 ]