+++ src/reactpy_django/router/__init__.py
@@ -1,4 +1,4 @@
-from reactpy_router.core import create_router
+from reactpy_router import create_router
 
 from reactpy_django.router.resolvers import DjangoResolver