+++ tests/test_app/apps.py
@@ -2,8 +2,8 @@ import contextlib
 import sys
 
 from django.apps import AppConfig
-from reactpy_django.utils import register_iframe
 
+from reactpy_django.utils import register_iframe
 from test_app import views
 
 
@@ -13,11 +13,11 @@ class TestAppConfig(AppConfig):
     def ready(self):
         from django.contrib.auth.models import User
 
-        register_iframe("test_app.views.view_to_component_sync_func_compatibility")
-        register_iframe(views.view_to_component_async_func_compatibility)
-        register_iframe(views.ViewToComponentSyncClassCompatibility)
-        register_iframe(views.ViewToComponentAsyncClassCompatibility)
-        register_iframe(views.ViewToComponentTemplateViewClassCompatibility)
+        register_iframe("test_app.views.view_to_iframe_sync_func")
+        register_iframe(views.view_to_iframe_async_func)
+        register_iframe(views.ViewToIframeSyncClass)
+        register_iframe(views.ViewToIframeAsyncClass)
+        register_iframe(views.ViewToIframeTemplateViewClass)
         register_iframe(views.view_to_iframe_args)
 
         if "test" in sys.argv: