+++ CHANGELOG.md
@@ -28,7 +28,9 @@ Using the following categories, list your changes in this order:
 
 ### Security
 -   for vulnerability fixes.
- -->
+
+Don't forget to remove deprecated code on each major release!
+-->
 
 <!--changelog-start-->